Finished profiel page

+++ mobile/src/pages/Profile/index.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useCallback, useEffect } from 'react';
 import { View, Text, ScrollView, Image, TouchableOpacity, KeyboardAvoidingView, Platform, Alert, ImageBackground } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
 import * as Yup from 'yup';
 import * as ImagePicker from 'expo-image-picker';
 
@@ -46,7 +45,6 @@ interface ProfileInfo {
 }
 
 const Profile: React.FC = () => {
-  const navigation = useNavigation();
   const { user, updateUser } = useAuth();
 
   const [profileInfo, setProfileInfo] = useState<ProfileInfo>({
@@ -83,6 +81,8 @@ const Profile: React.FC = () => {
   const [subject, setSubject] = useState('');
   const [cost, setCost] = useState('');
 
+  const [scheduleItems, setScheduleItems] = useState<ScheduleItem[] | null>(null);
+
   //Carregar todas as informações do usuário
   useEffect(() => {
     api.get<ProfileInfo>('logged-user').then(response => setProfileInfo(response.data));
@@ -98,42 +98,49 @@ const Profile: React.FC = () => {
 
     const formattedSchedule = profileInfo.class_schedule.map(scheduleItem => {
       return {
-        ...scheduleItem,
+        id: scheduleItem.id,
+        week_day: Number(scheduleItem.week_day),
         from: convertMinutesToHour(Number(scheduleItem.from)),
         to: convertMinutesToHour(Number(scheduleItem.to)),
       }
-    })
+    });
 
     setScheduleItems(formattedSchedule);
   }, [profileInfo])
 
-  const [scheduleItems, setScheduleItems] = useState<ScheduleItem[]>([
-    { id: 0, week_day: 0, from: '', to: '' }
-  ]);
-
   const addNewScheduleItem = useCallback(() => {
-    setScheduleItems([
-      ...scheduleItems,
-      { id: 0, week_day: 1, from: '', to: '' }
-    ])
+    if (scheduleItems) {
+      setScheduleItems([
+        ...scheduleItems,
+        { id: 0, week_day: 1, from: '', to: '' }
+      ]);
+    } else {
+      setScheduleItems([
+        { id: 0, week_day: 1, from: '', to: '' }
+      ]);
+    }
   }, [scheduleItems]);
 
   const removeScheduleItem = useCallback(async (id: number) => {
-    await api.delete(`/schedule-item/${id}`);
+    if (scheduleItems) {
+      await api.delete(`/schedule-item/${id}`);
 
-    const newScheduleItems = scheduleItems.filter(scheduleItem => scheduleItem.id !== id);
+      const newScheduleItems = scheduleItems.filter(scheduleItem => scheduleItem.id !== id);
 
-    setScheduleItems(newScheduleItems);
+      setScheduleItems(newScheduleItems);
+    }
   }, [scheduleItems])
 
   const setScheduleItemValue = useCallback((position: number, field: string, value: string) => {
-    const updateScheduleItems = scheduleItems.map((scheduleItem, index) => {
-      if (index === position) {
-        return { ...scheduleItem, [field]: value };
-      }
-      return scheduleItem;
-    });
-    setScheduleItems(updateScheduleItems);
+    if (scheduleItems) {
+      const updateScheduleItems = scheduleItems.map((scheduleItem, index) => {
+        if (index === position) {
+          return { ...scheduleItem, [field]: value };
+        }
+        return scheduleItem;
+      });
+      setScheduleItems(updateScheduleItems);
+    }
   }, [scheduleItems]);
 
   const handleUpdateProfile = useCallback(async () => {
@@ -185,7 +192,7 @@ const Profile: React.FC = () => {
       Alert.alert('Perfil atualizado com sucesso!');
 
     } catch (err) {
-      Alert.alert(err.message);
+      Alert.alert('Erro ao atualizar perfil!');
     }
   }, [name, email, whatsapp, bio, subject, cost, scheduleItems, updateUser, profileInfo.user.id, profileInfo.user.avatar, profileInfo.user.avatar_url]);
 
@@ -199,7 +206,7 @@ const Profile: React.FC = () => {
     }
 
     const pickerResult = await ImagePicker.launchImageLibraryAsync();
-    
+
     if (pickerResult.cancelled === true) {
       return;
     }
@@ -240,7 +247,7 @@ const Profile: React.FC = () => {
               </TouchableOpacity>
             </View>
 
-            <Text style={styles.userName} >{name}</Text>
+            <Text style={styles.userName} >{user.name}</Text>
             <Text style={styles.userSubject} >{subject}</Text>
           </View>
         </ImageBackground>
@@ -332,106 +339,105 @@ const Profile: React.FC = () => {
               </TouchableOpacity>
             </View>
 
-            {/* {scheduleItems.map((scheduleItem, index) => (
-              <View key={index} style={styles.scheduleItem}>
-                <Select
-                  label="Dia da semana"
-                  labelStyles={styles.labelStyle}
-                  items={[
-                    { label: 'Segunda', value: 1 },
-                    { label: 'Terça', value: 2 },
-                    { label: 'Quarta', value: 3 },
-                    { label: 'Quinta', value: 4 },
-                    { label: 'Sexta', value: 5 },
-                  ]}
-                  placeholder="Selecione"
-                  defaultValue={scheduleItem.week_day}
-                  onChangeItem={item => setScheduleItemValue(index, 'week_day', item.value)}
-                />
-
-                <View style={styles.scheduleItemTime}>
-                  <View style={styles.timeItem}>
-                    <Select
-                      label="Das"
-                      labelStyles={styles.labelStyle}
-                      items={[
-                        { label: '00h', value: "00:00" },
-                        { label: '01h', value: "01:00" },
-                        { label: '02h', value: "02:00" },
-                        { label: '03h', value: "03:00" },
-                        { label: '05h', value: "05:00" },
-                        { label: '06h', value: "06:00" },
-                        { label: '07h', value: "07:00" },
-                        { label: '08h', value: "08:00" },
-                        { label: '09h', value: "09:00" },
-                        { label: '10h', value: "10:00" },
-                        { label: '11h', value: "11:00" },
-                        { label: '12h', value: "12:00" },
-                        { label: '13h', value: "13:00" },
-                        { label: '14h', value: "14:00" },
-                        { label: '15h', value: "15:00" },
-                        { label: '16h', value: "16:00" },
-                        { label: '17h', value: "17:00" },
-                        { label: '18h', value: "18:00" },
-                        { label: '19h', value: "19:00" },
-                        { label: '20h', value: "20:00" },
-                        { label: '21h', value: "21:00" },
-                        { label: '22h', value: "22:00" },
-                        { label: '23h', value: "23:00" },
-                      ]}
-                      placeholder="Selecione"
-                      defaultValue={scheduleItem.from}
-                      onChangeItem={item => setScheduleItemValue(index, 'from', item.value)}
-                    />
+            {scheduleItems && scheduleItems.map((scheduleItem, index) => (
+                <View key={index} style={styles.scheduleItem}>
+                  <Select
+                    label="Dia da semana"
+                    labelStyles={styles.labelStyle}
+                    items={[
+                      { label: 'Segunda', value: 1 },
+                      { label: 'Terça', value: 2 },
+                      { label: 'Quarta', value: 3 },
+                      { label: 'Quinta', value: 4 },
+                      { label: 'Sexta', value: 5 },
+                    ]}
+                    placeholder="Selecione"
+                    defaultValue={scheduleItem.week_day ? scheduleItem.week_day : 1}
+                    onChangeItem={item => setScheduleItemValue(index, 'week_day', item.value)}
+                  />
+  
+                  <View style={styles.scheduleItemTime}>
+                    <View style={styles.timeItem}>
+                      <Select
+                        label="Das"
+                        labelStyles={styles.labelStyle}
+                        items={[
+                          { label: '00h', value: "00:00" },
+                          { label: '01h', value: "01:00" },
+                          { label: '02h', value: "02:00" },
+                          { label: '03h', value: "03:00" },
+                          { label: '05h', value: "05:00" },
+                          { label: '06h', value: "06:00" },
+                          { label: '07h', value: "07:00" },
+                          { label: '08h', value: "08:00" },
+                          { label: '09h', value: "09:00" },
+                          { label: '10h', value: "10:00" },
+                          { label: '11h', value: "11:00" },
+                          { label: '12h', value: "12:00" },
+                          { label: '13h', value: "13:00" },
+                          { label: '14h', value: "14:00" },
+                          { label: '15h', value: "15:00" },
+                          { label: '16h', value: "16:00" },
+                          { label: '17h', value: "17:00" },
+                          { label: '18h', value: "18:00" },
+                          { label: '19h', value: "19:00" },
+                          { label: '20h', value: "20:00" },
+                          { label: '21h', value: "21:00" },
+                          { label: '22h', value: "22:00" },
+                          { label: '23h', value: "23:00" },
+                        ]}
+                        placeholder="Selecione"
+                        defaultValue={scheduleItem.from}
+                        onChangeItem={item => setScheduleItemValue(index, 'from', item.value)}
+                      />
+                    </View>
+                    <View style={styles.timeItem}>
+                      <Select
+                        label="Até"
+                        labelStyles={styles.labelStyle}
+                        items={[
+                          { label: '00h', value: "00:00" },
+                          { label: '01h', value: "01:00" },
+                          { label: '02h', value: "02:00" },
+                          { label: '03h', value: "03:00" },
+                          { label: '05h', value: "05:00" },
+                          { label: '06h', value: "06:00" },
+                          { label: '07h', value: "07:00" },
+                          { label: '08h', value: "08:00" },
+                          { label: '09h', value: "09:00" },
+                          { label: '10h', value: "10:00" },
+                          { label: '11h', value: "11:00" },
+                          { label: '12h', value: "12:00" },
+                          { label: '13h', value: "13:00" },
+                          { label: '14h', value: "14:00" },
+                          { label: '15h', value: "15:00" },
+                          { label: '16h', value: "16:00" },
+                          { label: '17h', value: "17:00" },
+                          { label: '18h', value: "18:00" },
+                          { label: '19h', value: "19:00" },
+                          { label: '20h', value: "20:00" },
+                          { label: '21h', value: "21:00" },
+                          { label: '22h', value: "22:00" },
+                          { label: '23h', value: "23:00" },
+                        ]}
+                        placeholder="Selecione"
+                        defaultValue={scheduleItem.to}
+                        onChangeItem={item => setScheduleItemValue(index, 'to', item.value)}
+                      />
+                    </View>
                   </View>
-                  <View style={styles.timeItem}>
-                    <Select
-                      label="Até"
-                      labelStyles={styles.labelStyle}
-                      items={[
-                        { label: '00h', value: "00:00" },
-                        { label: '01h', value: "01:00" },
-                        { label: '02h', value: "02:00" },
-                        { label: '03h', value: "03:00" },
-                        { label: '05h', value: "05:00" },
-                        { label: '06h', value: "06:00" },
-                        { label: '07h', value: "07:00" },
-                        { label: '08h', value: "08:00" },
-                        { label: '09h', value: "09:00" },
-                        { label: '10h', value: "10:00" },
-                        { label: '11h', value: "11:00" },
-                        { label: '12h', value: "12:00" },
-                        { label: '13h', value: "13:00" },
-                        { label: '14h', value: "14:00" },
-                        { label: '15h', value: "15:00" },
-                        { label: '16h', value: "16:00" },
-                        { label: '17h', value: "17:00" },
-                        { label: '18h', value: "18:00" },
-                        { label: '19h', value: "19:00" },
-                        { label: '20h', value: "20:00" },
-                        { label: '21h', value: "21:00" },
-                        { label: '22h', value: "22:00" },
-                        { label: '23h', value: "23:00" },
-                      ]}
-                      placeholder="Selecione"
-                      defaultValue={scheduleItem.to}
-                      onChangeItem={item => setScheduleItemValue(index, 'to', item.value)}
-                    />
+                  <View style={styles.removeButtonContainer}>
+                    <View style={styles.line} />
+                    <TouchableOpacity
+                      style={styles.removeButton}
+                      onPress={() => { removeScheduleItem(scheduleItem.id) }}
+                    >
+                      <Text style={styles.removeButtonText}>Excluir horário</Text>
+                    </TouchableOpacity>
+                    <View style={styles.line} />
                   </View>
                 </View>
-                <View style={styles.removeButtonContainer}>
-                  <View style={styles.line} />
-                  <TouchableOpacity 
-                    style={styles.removeButton}
-                    onPress={() => {removeScheduleItem(scheduleItem.id)}}
-                  >
-                    <Text style={styles.removeButtonText}>Excluir horário</Text>
-                  </TouchableOpacity>
-                  <View style={styles.line} />
-                </View>
-              </View>
-            ))} */}
-
+              ))}
 
           </View>
           <View style={styles.giveClassesFormFooter}>
